+++ client-app/src/components/ui/breadcrumb.tsx
@@ -1,8 +1,7 @@
 import * as React from "react"
 import { Slot } from "@radix-ui/react-slot"
 import { ChevronRight, MoreHorizontal } from "lucide-react"
-
-import { cn } from "@/lib/utils"
+import { cn } from "@/helpers/concat-class-names.helper"
 
 const Breadcrumb = React.forwardRef<
   HTMLElement,
