+++ client-app/src/components/ui/card.tsx
@@ -1,6 +1,5 @@
 import * as React from "react"
-
-import { cn } from "@/lib/utils"
+import { cn } from "@/helpers/concat-class-names.helper"
 
 const Card = React.forwardRef<
   HTMLDivElement,
