+++ client-app/src/components/ui/dropdown-menu.tsx
@@ -3,8 +3,7 @@
 import * as React from "react"
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu"
 import { Check, ChevronRight, Circle } from "lucide-react"
-
-import { cn } from "@/lib/utils"
+import { cn } from "@/helpers/concat-class-names.helper"
 
 const DropdownMenu = DropdownMenuPrimitive.Root
