+++ client-app/src/components/ui/popover.tsx
@@ -1,7 +1,6 @@
 import * as React from "react"
 import * as PopoverPrimitive from "@radix-ui/react-popover"
-
-import { cn } from "@/lib/utils"
+import { cn } from "@/helpers/concat-class-names.helper"
 
 const Popover = PopoverPrimitive.Root
