+++ client-app/src/components/ui/separator.tsx
@@ -1,7 +1,6 @@
 import * as React from "react"
 import * as SeparatorPrimitive from "@radix-ui/react-separator"
-
-import { cn } from "@/lib/utils"
+import { cn } from "@/helpers/concat-class-names.helper"
 
 const Separator = React.forwardRef<
   React.ElementRef<typeof SeparatorPrimitive.Root>,
