+++ client-app/src/components/ui/tooltip.tsx
@@ -1,7 +1,6 @@
 import * as React from "react"
 import * as TooltipPrimitive from "@radix-ui/react-tooltip"
-
-import { cn } from "@/lib/utils"
+import { cn } from "@/helpers/concat-class-names.helper"
 
 const TooltipProvider = TooltipPrimitive.Provider
