+++ client-app/src/helpers/concat-class-names.helper.ts
@@ -1,0 +1,8 @@
+import clsx, { ClassValue } from "clsx"
+import { twMerge } from "tailwind-merge"
+
+export function concatClassNames(...inputs: ClassValue[]) {
+    return twMerge(clsx(inputs))
+}
+
+export const cn = concatClassNames
