+++ client-app/src/helpers/convert-base64-to-blob.helper.ts
@@ -1,0 +1,11 @@
+export function convertBase64ToBlob(base64: string) {
+    const arr = base64.split(",")
+    const mime = arr[0].match(/:(.*?);/)![1]
+    const bstr = atob(arr[1])
+    let n = bstr.length
+    const u8arr = new Uint8Array(n)
+    while (n--) {
+        u8arr[n] = bstr.charCodeAt(n)
+    }
+    return new Blob([u8arr], { type: mime })
+}
