--- client-app/src/lib/constants.ts
@@ -1,1 +1,0 @@
-export const CONSTANT_A = "CONSTANT_A"
